filters and entry standard of issue
Filters for user, dates or other fields are added, additionally you can pass the orderBy parameter to sort in ascending and descending order.
The way to create the issues is standardized, being only necessary to pass the title and description, since the other data is generated in code and the user is obtained from the authentication validation.